requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
anthropic==0.18.1
requests==2.31.0
playwright==1.40.0
